refactor: Move IHandle<> to Microsoft.Autogen.Contracts
