+++ dotnet/samples/dev-team/DevTeam.Backend/Agents/AzureGenie.cs
@@ -2,7 +2,9 @@
 // AzureGenie.cs
 
 using DevTeam.Backend.Services;
+using Microsoft.AutoGen.Contracts;
 using Microsoft.AutoGen.Core;
+
 namespace DevTeam.Backend.Agents;
 
 [TopicSubscription(Consts.TopicName)]
