+++ dotnet/samples/dev-team/DevTeam.Backend/Agents/Developer/Developer.cs
@@ -2,6 +2,7 @@
 // Developer.cs
 
 using DevTeam.Agents;
+using Microsoft.AutoGen.Contracts;
 using Microsoft.AutoGen.Core;
 
 namespace DevTeam.Backend.Agents.Developer;
