+++ dotnet/samples/dev-team/DevTeam.Backend/Agents/DeveloperLead/DeveloperLead.cs
@@ -2,6 +2,7 @@
 // DeveloperLead.cs
 
 using DevTeam.Agents;
+using Microsoft.AutoGen.Contracts;
 using Microsoft.AutoGen.Core;
 
 namespace DevTeam.Backend.Agents.DeveloperLead;
