+++ dotnet/samples/dev-team/DevTeam.Backend/Agents/ProductManager/ProductManager.cs
@@ -2,6 +2,7 @@
 // ProductManager.cs
 
 using DevTeam.Agents;
+using Microsoft.AutoGen.Contracts;
 using Microsoft.AutoGen.Core;
 
 namespace DevTeam.Backend.Agents.ProductManager;
