+++ dotnet/src/Microsoft.AutoGen/Core/HandlerInvoker.cs
@@ -3,6 +3,7 @@
 
 using System.Diagnostics;
 using System.Reflection;
+
 using Microsoft.AutoGen.Contracts;
 
 namespace Microsoft.AutoGen.Core;
