+++ dotnet/test/Microsoft.AutoGen.Runtime.Grpc.Tests/TestAgent.cs
@@ -2,6 +2,7 @@
 // TestAgent.cs
 
 using System.Collections.Concurrent;
+using Microsoft.AutoGen.Contracts;
 using Microsoft.AutoGen.Core;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
